+++ src/mpfrmatcoal.h
@@ -6,8 +6,8 @@
 int  MpfrMatCoal_isInitialized(void);
 void MpfrMatCoal_initExterns(long nsamp);
 void MpfrMatCoal_freeExterns(void);
-void MpfrMatCoal_project(int dim, double ans[dim], double v);
-void MpfrMatCoal_ciLen(int dim, double ans[dim], double v);
+void MpfrMatCoal_project(int dim, long double ans[dim], long double v);
+void MpfrMatCoal_ciLen(int dim, long double ans[dim], long double v);
 void MpfrMatCoal_printAll(FILE *fp);
 
 #endif
